Shortened the length for edges for glycolysis
Need to make physical model shorter so that changing the camera angle does not interfere with rendering certain meshes on top of each other.
Meshes are not opaque so it doesn't use depth buffer.
They are transparent so they are rendered differently (apparently based on the order of how far the pivot is away from the camera?).